Server switching now wraps around
git-svn-id: http://jules.aerix.nl/svn/g-key/trunk@206 f5aa9b1c-f3d2-43f6-b11e-4419a1f85e2c

+++ plugin.cpp
@@ -440,14 +440,14 @@ void ParseCommand(char* cmd, char* arg)
 	}
 	else if(!strcmp(cmd, "TS3_VOLUME_UP"))
 	{
-		float diff = (arg!=NULL && *arg != (char)NULL)?atof(arg):1.0f;
+		float diff = (arg!=NULL && *arg != (char)NULL)?(float)atof(arg):1.0f;
 		float value;
 		ts3Functions.getPlaybackConfigValueAsFloat(scHandlerID, "volume_modifier", &value);
 		SetMasterVolume(scHandlerID, value+diff);
 	}
 	else if(!strcmp(cmd, "TS3_VOLUME_DOWN"))
 	{
-		float diff = (arg!=NULL && *arg != (char)NULL)?atof(arg):1.0f;
+		float diff = (arg!=NULL && *arg != (char)NULL)?(float)atof(arg):1.0f;
 		float value;
 		ts3Functions.getPlaybackConfigValueAsFloat(scHandlerID, "volume_modifier", &value);
 		SetMasterVolume(scHandlerID, value-diff);
